// src/app/models/city-weather.model.ts
export class CityWeather {
    name: string;
    date: Date;
    pressure: number;
    humidity: number;
    temp: number;
    tempMax: number;
    tempMin: number;
    lat: number;
    lon: number;
    country: string;
    sunrise: Date;
    sunset: Date;
    windSpeed: number;
    windDegree: number;
    description: string;
}